Move isResultCompatibleWithDisposal function into root lib folder
