+++ packages/core/lib/isResultCompatibleWithDisposal.test.ts
@@ -1,6 +1,6 @@
-import type { Offence } from "../../types/AnnotatedHearingOutcome"
+import type { Offence } from "../types/AnnotatedHearingOutcome"
 
-import ResultClass from "../../types/ResultClass"
+import ResultClass from "../types/ResultClass"
 import isResultCompatibleWithDisposal from "./isResultCompatibleWithDisposal"
 
 describe("isResultCompatibleWithDisposal", () => {
