+++ packages/core/lib/triggers/TRPS0010.test.ts
@@ -1,15 +1,15 @@
 jest.mock("../../phase2/lib/isRecordableOffence")
 jest.mock("./hasCompletedDisposal")
-jest.mock("../../phase2/lib/isResultCompatibleWithDisposal")
+jest.mock("../isResultCompatibleWithDisposal")
 
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 
 import isRecordableOffence from "../../phase2/lib/isRecordableOffence"
-import isResultCompatibleWithDisposal from "../../phase2/lib/isResultCompatibleWithDisposal"
 import generateAhoFromOffenceList from "../../phase2/tests/fixtures/helpers/generateAhoFromOffenceList"
 import generatePncUpdateDatasetFromOffenceList from "../../phase2/tests/fixtures/helpers/generatePncUpdateDatasetFromOffenceList"
 import Phase from "../../types/Phase"
 import { PncOperation } from "../../types/PncOperation"
+import isResultCompatibleWithDisposal from "../isResultCompatibleWithDisposal"
 import hasCompletedDisposal from "./hasCompletedDisposal"
 import TRPS0010 from "./TRPS0010"
