+++ packages/core/lib/triggers/TRPS0010.ts
@@ -3,9 +3,9 @@ import TriggerCode from "@moj-bichard7-developers/bichard7-next-data/dist/types/
 import type { TriggerGenerator } from "../../types/TriggerGenerator"
 
 import isRecordableOffence from "../../phase2/lib/isRecordableOffence"
-import isResultCompatibleWithDisposal from "../../phase2/lib/isResultCompatibleWithDisposal"
 import Phase from "../../types/Phase"
 import { isPncUpdateDataset } from "../../types/PncUpdateDataset"
+import isResultCompatibleWithDisposal from "../isResultCompatibleWithDisposal"
 import hasCompletedDisposal from "./hasCompletedDisposal"
 
 const triggerCode = TriggerCode.TRPS0010
