+++ packages/core/lib/triggers/TRPS0011.test.ts
@@ -1,16 +1,16 @@
 import type { Offence } from "../../types/AnnotatedHearingOutcome"
 
 import isRecordableOffence from "../../phase2/lib/isRecordableOffence"
-import isResultCompatibleWithDisposal from "../../phase2/lib/isResultCompatibleWithDisposal"
 import generateAhoFromOffenceList from "../../phase2/tests/fixtures/helpers/generateAhoFromOffenceList"
 import generatePncUpdateDatasetFromOffenceList from "../../phase2/tests/fixtures/helpers/generatePncUpdateDatasetFromOffenceList"
 import Phase from "../../types/Phase"
 import { PncOperation } from "../../types/PncOperation"
+import isResultCompatibleWithDisposal from "../isResultCompatibleWithDisposal"
 import hasCompletedDisposal from "./hasCompletedDisposal"
 import TRPS0011 from "./TRPS0011"
 jest.mock("../../phase2/lib/isRecordableOffence")
 jest.mock("./hasCompletedDisposal")
-jest.mock("../../phase2/lib/isResultCompatibleWithDisposal")
+jest.mock("../isResultCompatibleWithDisposal")
 
 const mockedIsRecordableOffence = isRecordableOffence as jest.Mock
 const mockedHasCompletedDisarr = hasCompletedDisposal as jest.Mock
