+++ packages/core/phase3/lib/hearingDetails/generateArrestHearingsAdjudicationsAndDisposals.ts
@@ -2,7 +2,7 @@ import type { PncUpdateDataset } from "../../../types/PncUpdateDataset"
 import type { PncUpdateArrestHearingAdjudicationAndDisposal } from "../../types/HearingDetails"
 
 import getAdjustedRecordableOffencesForCourtCase from "../../../lib/getAdjustedRecordableOffencesForCourtCase"
-import isResultCompatibleWithDisposal from "../../../phase2/lib/isResultCompatibleWithDisposal"
+import isResultCompatibleWithDisposal from "../../../lib/isResultCompatibleWithDisposal"
 import { createAdjudicationFromOffence } from "./createAdjudicationFromOffence"
 import { createArrestHearingFromOffence } from "./createArrestHearingFromOffence"
 import createDisposalsFromOffence from "./createDisposalsFromOffence"
